pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "mixxxdb"
version = "0.0.1"
dependencies = ["python-dateutil"]

[project.optional-dependencies]
dev = [
    "mypy",
    "ipython",
    "ipdb",
    "isort",
    "types-python-dateutil",
]

[project.scripts]
mixxxdb-mv = "mixxxdb.cli:mv"
mixxxdb-orphanfiles = "mixxxdb.cli:orphanfiles"
mixxxdb-orphanlocations = "mixxxdb.cli:orphanlocations"

[tool.isort]
line_length = 119
multi_line_output = 3
combine_as_imports = true
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "SELF", "LOCALFOLDER"]
known_self = "mixxxdb"
lines_after_imports = 2
include_trailing_comma = true
skip = [".venv"]
